refactoring of image canvas into hooks and subcomponents
