+++ package.json
@@ -17,9 +17,10 @@
     "react-remove-scroll": "^2.0.4",
     "react-select": "^2.2.0",
     "react-syntax-highlighter": "^12.2.1",
-    "react-use": "^13.26.3",
+    "react-use": "^13.27.0",
     "seamless-immutable": "^7.1.4",
     "transformation-matrix-js": "^2.7.6",
+    "use-event-callback": "^0.1.0",
     "use-key-hook": "^1.3.0"
   },
   "peerDependencies": {
